+++ Configuration/TCA/tx_campuseventsconnector_domain_model_sponsor.php
@@ -173,6 +173,8 @@ return [
                     'custom_process' => 'files', //Field handled separately
                     'size_field' => ['imageFile', 'size'],
                     'alt_text_source_field' => 'name',
+                    'timestamp_import_field' => ['imageFile', 'modifiedAt'],
+                    'timestamp_normalizer' => 'datetime_to_tstamp',
                 ],
             ],
         ]